// client/api-client/models/create-rule-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Vault API
 * Vault API
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface CreateRuleDTO
 */
export interface CreateRuleDTO {
    /**
     * 
     * @type {string}
     * @memberof CreateRuleDTO
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof CreateRuleDTO
     */
    criteria: string;
    /**
     * 
     * @type {boolean}
     * @memberof CreateRuleDTO
     */
    isEnabled: boolean;
    /**
     * 
     * @type {string}
     * @memberof CreateRuleDTO
     */
    field: CreateRuleDTOFieldEnum;
    /**
     * 
     * @type {string}
     * @memberof CreateRuleDTO
     */
    operation: CreateRuleDTOOperationEnum;
}

/**
    * @export
    * @enum {string}
    */
export enum CreateRuleDTOFieldEnum {
    Name = 'name',
    TransactionId = 'transactionId',
    PaymentChannel = 'paymentChannel'
}
/**
    * @export
    * @enum {string}
    */
export enum CreateRuleDTOOperationEnum {
    Contains = 'contains',
    Equals = 'equals'
}